# openapi.yaml
openapi: 3.0.3
info:
  title: Babbel API
  description: |
    HTTP API for generating audio news bulletins. Combines news stories with station-specific jingles using FFmpeg.
    
    ## Authentication
    
    - Local: Username/password with session cookies
    - OAuth/OIDC: Azure AD, Google, Okta support
    - Account lockout protection
    
    ## Authorization
    
    Role-based access control:
    - admin: Full access including user management
    - editor: Content management (stations, voices, stories, bulletins)
    - viewer: Read-only access
    
    ## Architecture
    
    Station-voice junction table:
    - Each voice has different jingles per station
    - Mix point defines when voice starts over jingle
    - Files stored as `station_{id}_voice_{id}_jingle.wav`
  version: 1.0.0
  x-changelog:
    1.0.0:
      - "Initial stable release"
      - "HTTP API for audio news bulletin generation"
      - "Station-specific voice jingles with mix points"
      - "Local and OAuth/OIDC authentication"
      - "Role-based access control (admin, editor, viewer)"
      - "Soft delete for stories and users"
      - "FFmpeg audio processing"
      - "Pagination and filtering on all list endpoints"
      - "RESTful design for radio automation"
      - "Removed redundant /broadcasts endpoint"
  contact:
    name: Streekomroep ZuidWest
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://your-api-server.com/api/v1
...
security:
  - cookieAuth: []
paths:
...
  /stories/{id}:
    get:
      summary: Get story by ID
      tags:
        - Stories
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Story details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update story
      tags:
        - Stories
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StoryInput'
      responses:
        '200':
          description: Story updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Story'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete story (soft delete)
      description: |
        Performs a soft delete by setting deleted_at timestamp. 
        The story will no longer appear in listings unless include_deleted=true is specified.
        Story data and audio files are preserved for potential restoration.
      tags:
        - Stories
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204':
          description: Story soft deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update story state
      description: Update story status (draft, active, expired) or other state fields without requiring full story data.
      tags:
        - Stories
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [draft, active, expired]
                  description: Story status
              example:
                status: "active"
      responses:
        '200':
          description: Story state updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Story status updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: babbel_session
      description: Session-based authentication using secure cookies
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Resource ID
...
  schemas:
...
    Voice:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Story:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        text:
          type: string
        voice_id:
          type: integer
        voice:
          $ref: '#/components/schemas/Voice'
        audio_url:
          type: string
          nullable: true
          description: API URL to download the audio file
        duration_seconds:
          type: number
          format: float
          nullable: true
        status:
          type: string
          enum: [draft, active, expired]
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        weekdays:
          type: object
          properties:
            monday:
              type: boolean
            tuesday:
              type: boolean
            wednesday:
              type: boolean
            thursday:
              type: boolean
            friday:
              type: boolean
            saturday:
              type: boolean
            sunday:
              type: boolean
          description: |
            Weekday selection for scheduling. 
            API returns JSON object format for convenience.
            Internally stored as bitmask for efficiency.
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the story was soft-deleted
    StoryInput:
      type: object
      required:
        - title
        - text
        - voice_id
        - start_date
        - end_date
      properties:
        title:
          type: string
        text:
          type: string
        voice_id:
          type: integer
        audio:
          type: string
          format: binary
          description: Audio file (optional)
        status:
          type: string
          enum: [draft, active, expired]
          default: draft
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        weekdays:
          type: object
          properties:
            monday:
              type: boolean
              default: false
            tuesday:
              type: boolean
              default: false
            wednesday:
              type: boolean
              default: false
            thursday:
              type: boolean
              default: false
            friday:
              type: boolean
              default: false
            saturday:
              type: boolean
              default: false
            sunday:
              type: boolean
              default: false
          description: |
            Weekday selection for scheduling. Pass as JSON object in multipart form.
            Example: {"monday":true,"tuesday":true,"wednesday":true,"thursday":true,"friday":true,"saturday":false,"sunday":false}
            Internally stored as bitmask but API accepts/returns JSON object format.
        metadata:
          type: object
          additionalProperties: true
          default: {}
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...